Split AttestationService out into an interface and an impl per the team standards.
